uniontraits: check for Union type via a concrete instance
Previously we performed tests against Union symbol itself to check
for inheritance.

This relied on an implementation detail where an instance of a generic
in the AST uses the same symbol as the generic type.

The PR https://github.com/nim-lang/Nim/pull/22642 changed this
implementation detail. So now, we have to account for both cases of new
and old compilers by also testing against a pre-defined instance.

Fixes https://github.com/alaviss/union/issues/51